requirements.txt
flask
laserembeddings
numpy
torch
ufal.udpipe
sklearn
pyyaml
requests
flask-markdown
